add user modified

+++ src/main/java/com/example/login_logout_30/CustomUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface CustomUserRepository extends JpaRepository<CustomUser, Integer> {
+public interface CustomUserRepository extends JpaRepository<CustomUser, Long> {
     @Query("select c from CustomUser c where c.name = ?1")
     Optional<CustomUser> findByNameEquals(String name);
